phpunit: Speed up MediaWikiCoversValidator trait
Using a data provider for a dynamic or otherwise potentially large amount
of cases tends to come with significant overhead due to all the setup
and test abstraction involved.

Especially when the host class has expensive setUp() or tearDown(),
logic such as the case for API integration tests and other test cases
that have `group Database`. Each of the test methods invokes the
DB reset logic for no good reason.

As an example, ApiQueryRecentChangesIntegrationTest takes about
55-60 seconds to run locally for me on master, but takes only 35-40s
after this patch.

Before:

> slow tests (>50ms):
>  …
>  44. 107ms to run ApiQueryRecentChangesIn…t:testValidCovers with d… #25
>  45. 107ms to run ApiQueryRecentChangesIn…t:testValidCovers with d… #2
>  46. 107ms to run ApiQueryRecentChangesIn…t:testValidCovers with d… #7
>  47. 107ms to run ApiQueryRecentChangesIn…t:testValidCovers with d… #8
>  48. 107ms to run ApiQueryRecentChangesIn…t:testValidCovers with d… #14
>  49. 107ms to run ApiQueryRecentChangesIn…t:testValidCovers with d… #17
>  50. 106ms to run ApiQueryRecentChangesIn…t:testValidCovers with d… #18
> ...and there are 3 more above your threshold hidden from view
>
> Time: 59.13 seconds
> OK (53 tests)

After:

> slow tests (>50ms):
> …
> 26. 122ms to run ApiQueryRecentChangesIntegrationTest:testValidCovers
> 27. 108ms to run ApiQueryRecentChangesIntegrationTest:testApiTestGroup
>
> Time: 40.8 seconds
> OK (27 tests)

And that's just for running ApiQueryRecentChangesIntegrationTest.php
in isolation. I expect the CI jobs overall to come down by much more
than that.

Bug: T225730
Change-Id: Idae95f7c6f44c02371e2abc6e41d79b9eaced425