Add expiry select field to action=watch
If $wgWatchlistExpiry is enabled, add a new form field to
the watch page at ?action=watch. Display a different form
header depending on whether the page is already being watched
(either permanently or temporarily). The existing behaviour
does not change (including not showing any indication that
a page is already being watched).

Bug: T245565
Change-Id: I77362dfbb5f49b2657532eb43de2ffa80331d5cf