SiteConfiguration: Remove by-ref return from getLocalDatabases()
This is most likely a micro-optimisation from a time where MW
still ran mainly on PHP 4, during which copy-on-write wasn't as
well implemented (or not as well understood).

Nowadays, this actually makes the method slower, and also makes
the internals hard to reason about due to the side effects a
downstream mutation could cause.

Remove this undocumented feature, and also make the method
explicitly public.

Bug: T169821
Change-Id: I1887021282bbc4ab619a46e80a8c54bf1fc794ab