ExtensionRegistry: Avoid losing 'defines' when loading lazy-loaded attributes
When fetching lazy-loaded attributes in ExtensionRegistry to cache, we are saving all the values
from extension.json a second time. In doing so, we are wrongfully omitting the values
previously being defined for the "defines" attribute
(the attribute that is responsible for setting namespace constants).
This patch will make sure that we keep the value set in the "defines" attribute
so that the constants are defined when we load from cache the second time.

Bug: T245629
Change-Id: I4f151f88ece56cf718749b9de11fc8e204ccf29d