Allow users to set tabindex="0" on elements
Important for keyboard focusability of elements in order to ensure for 
example users with motoric impairments to reach those elements.
This patch does not allow setting tabindex="-1" or tabindex > 0.
tabindex > 1 seems like a terrible idea to allow users to do.
I don't see any valid reason for tabindex="-1" in wikitext, so
lets not allow that for now either.

Bug: T247910
Change-Id: I5065b2deeb14bdb3682dd176b87f254ac6f2cf88