build: Remove old ZUUL_PROJECT hack from Gruntfile.js
This was here from a year ago when we added CHROMIUM_FLAGS, which
has since rolled out.

Also make the assert-mw-env less fragile by checking the
env variables in one place each instead of two places. Before this,
it was relatively easy to accidentally add a warning for one, without
the shell exit code changing.

Change-Id: I6b85483981b415fe251b3760d5f282433e7fc4e9