Use errorbox/warningbox for all HTMLFormField errors
Continuing the work for 441e12f2d9c8a51935ae735ed51c48ed4e80bcc0,
which did this for the main HTMLForm errors (grouped at the top of
the form). This handles errors for individual fields.

This fixes the display of error/warning messages on Special:CreateAccount,
and should not break anything else too badly (it helps that most forms
are using OOUI these days).

Bug: T246894
Change-Id: If1ccdff10bebe0c2eeca4c15d82271c9de1af164