phpunit: Fix omission of 'integration/' in phpunit.php runner
Follows-up 214750d8d2 which created this directory, but forgot
to add it in tests/phpunit/suite.xml.

This was only realized this week, 9 months later, as part of
I29f5e761b977c where an integration test was fixed due to
observing a failure locally when running it directly, and this
not having been caught by CI when the failure was introduced,
because the suite is not being run.

Change-Id: I9edb8fad870ec4345c102e3cc2d280f872d20a22