rdbms: Don't silence errors in DatabaseMysqlBase::serverIsReadOnly()
It's highly unlikely that that query would ever error, but if it does
(and apparently it does sometimes somehow) we need to either handle the
error or let it be raised.

Since there doesn't seem to be any particularly sane thing to assume if
there is an error, let's go with "let it be raised".

Bug: T247865
Change-Id: I6c0e08df90eb46953ba5eb6b5a3e8c6f52929564