rcfilters: Fix focus order on the limit/duration popup
Narrator not announcing popup dialog when
when changes and limit button is clicked.

Focus order is not logical when for keyboard only users

This fix here attempts to make the focus logical, it binds a handler
to the ready and close event placing focus on the button select widget
and place focus back on the button when the popup closes.

Bug: T257171
Change-Id: I66355d5d4d45a9b4aeb14974d228e4a193645a1e