uppercaseTitlesForUnicodeTransition.php: Improve handling of non-moves
Doing the actual page moves first caused the entries in archive,
logging, and so on to either not be moved or to be put to the
prefixed/suffixed title, because the page move created the title and
then mungeTitle refused to use it again later. If we do archive,
logging, and so on first, that doesn't happen.

Also, shouldn't apply the prefix or suffix at all when processing the
redirect table (rd_title), as we're not changing the actual wikitext. We
should only uppercase the rd_title to match what MediaWiki would do if
the redirect were purged.

Bug: T219279
Change-Id: I2224ac7cb7e952c7cff5a576583d4cf1bc05e462