uppercaseTitlesForUnicodeTransition.php: Delete useless redirects
If the lowercase title is already a redirect to the uppercase title, or
the lowercase and uppercase titles are both redirects to the same
target, there's probably no point in renaming it.

Note we still do the rename to the prefixed/suffixed title before
deleting, so the resulting archive and logging rows wind up under the
correct title.

Bug: T219279
Change-Id: I6c5e3ab39a0e136b4a3f96e9c78d36eea636bc4d