phpunit: Consider "risky" tests as failed tests
Jenkins, or more precise, junit, already detects PHPUnit's risky tests.
This is currently mapped to the "UNSTABLE" build status for Jenkins jobs.
Our Zuul confguration does consider UNSTABLE as a failure, so it does
result in the appropiate V-1 score already.

But, "UNSTABLE" is confusing in the Gerrit comments from jenkins-bot,
but it is merely about a PHPUnit test, for example:
https://integration.wikimedia.org/ci/job/wmf-quibble-vendor-mysql-php72-docker/28970/

This patch lets phpunit fail loudly and explicitly on risky tests,
and thus makes them easier to find and fix.

Change-Id: I95a54d1e4c8bf6b6abcc698bc14c2db5ce6c2a7d