DatabaseUpdater: protect direct modification methods
Methods that directly modify the database should not be public.
So far, some methods where protected, some where public, in some
cases misleading LoadExtensionSchemaUpdates handlers to call them
directly, leading to dangerous behavior.

NOTE: this is a breaking change without deprecation. Extensions
that break due to this change need to be updated, since the
old behavior may lead to database corruption and data loss.

As of this writing, the following extensions erronously use one
of the affected methods:

* Wikibase: dropTable
* CreateWiki: modifyTable, modifyField
* ManageWiki: modifyTable
* MatomoAnalytics: modifyTable
* ArticleFeedbackv5: modifyField
* OpenIDConnect: modifyField
* MirahezeMagic: modifyField
* Wikifab: modifyField

Bug: T249603
Change-Id: Ie73e6143bb5cbce0dae705a7451fb49cf7e3abcd