registration: Convert internal static class vars to constants
Constants with visibility modifiers are new in PHP 7.1, and most
uses of protected/private static vars were essentially constants
in disguise. Now that we can recognise them as such, do so.

None of these variable names were found anywhere in repos indexed
by Codesearch, except for this in this very file.

Change-Id: I1abfd58ce3bed734f33676ffa7a0ce66ff790d8f