resourceloader: Use RevisionRecord in WikiModule::invalidateModuleCache
Replace uses of Revision with RevisionRecord; passing a Revision is not
supported anymore, and the method is now marked as @internal.

Decide whether to purge the cache based on the content models of the
revisions' main slots, rather than their formats. The content format
is more difficult to determine on a RevisionRecord than with Revision.
The new interface encourages use of content models instead, so we use
that from now on as well.

Bug: T249561
Bug: T250981
Change-Id: I7ad27f52c0975b0c22f2aa9248c3fcf6d5b2ed8f