Add `Monolog` to production-builds
When a package is build with the `composer update --no-dev` command the
`monolog/monolog` package will not be included. But for advanced
debugging it is very important to have this in a customer setup.

ATTENTION: This should be kept in snyc with the package requirement in
`<mediawiki>/composer.json/require-dev`

[3.1.x]