Show standard diff elements if revision is null only because of permission
Article::fetchRevisionRecord() returns revision record if it exists and the context user is allowed to view it. So it can be null either because it doesn't exist or lack of permission.

However Article::showDiffPage treats the null as just no revision all the time and that caused T213621.

In this patch, if we encounter that null, we'll try to load the revision from the id to check which null we're dealing with. If loading from id returns null again, that means it doesn't really exist. No standard diff page, and error message will be returned.

But if it returns RevisionRecord, then we know the null is because of permission. And as we proceed, the revision will be used to call Content Handler and DifferenceEngine classes which are responsible for showing the headers, error message as well as actually showing/hiding the revision content as approriatei

Bug: T213621
Change-Id: I703ecc6fbd3555d93ce34f819efdb953c38f4231