build: Document a one-line dependency-free pure php install method
This has been mentioned and referenced from time to time, but having
it defined in composer.json would make it much easier to refer to
and make use of without having to worry about the specific parameters.

This is essentially the same way that Quibble works currently.

Also, simplify the Travis CI setup by using this instead of
concerning that configuration with MySQL and Travis' specific
way of dealing with that.

Usage:

 # Install with pure PHP and start a server (no Apache/MySQL/..)
 $ composer mw-install
 $ composer serve
 $ open http://localhost:4000

Based on <https://wikitech.wikimedia.org/wiki/Fresnel#Quick_MediaWiki>.

Bug: T56856
Change-Id: Ib1f8a86afe48b8f1ae6db733735e81f51d698fc9