jquery.tablesorter: Allow top sorting via 'sorttop' class
Allow rows to be sorted to the top of a table by adding a 'sorttop'
class. This class is handled similarly to the 'sortbottom' class:
rows are appended to the <thead>, as 'sortbottom' rows are prepended
to the <tfoot>. (The sorting algorithm ignores header and footer
rows.)

Additionally, the sort buttons will not be appended to any row
with the 'sorttop' class.

Bug: T248936
Change-Id: I9fb307dd7266798a32c4e15bc8bf3e910efe8f38