Don't suppress deprecation log messages
Send deprecation messages to $wgDebugLogFile if it is configured.
Previously deprecation messages were marked as "private" and so were
shown only if someone specifically configured the channel in
$wgDebugLogGroups.

Complete suppression of deprecation messages was an accident stemming
from an attempt to avoid duplication of those messages when
$wgDebugToolbar is enabled. However, that duplication is harmless and
perhaps even useful -- we see an HTML deprecation message in the
console, and also a plain text deprecation message in its context in the
toolbar's debug log, which seems fine to me.

Change-Id: I8edb41aa1d8236047b80830dcdda546e326b09ec