Fix Postgres user_newtalk table
In mysql the user_id and user_ip fields are not nullable,
but include a default 0 and '', however in Postgress the default
was not set, while user_id was set to not nullable. In the
code that uses this table, we set either user_id or user_ip
depending on whether the user is anon or registered. Thus both
fields should either be nullable, or contain a default.
Given that mysql has defaults set, I've opted for bringing
Postgress closer to what mysql does.

This was always broken and was discovered by adding new tests
for this functionality.

Bug: T239640
Change-Id: I75bf469a30225687a4b0cb550a4068cb07208c01