Make Parser::$mVariables private
This was overlooked in a5afda55bfbdf7b18b21ad3b41ec281bf8f5130c, probably
because casual inspection of
 https://codesearch.wmflabs.org/deployed/?q=mVariables&i=nope&files=&repos=
made it look like extensions/AbuseFilter was using Parser::$mVariables.

But AbuseFilterParser (and AbuseFilterCachingParser) is its own thing, not
a subclass of Parser, and the (re)use of the $mVariables name for the
property is just coincidence.

Change-Id: I4debb11c804b2ef4f1b727d521f38f35de691708