maintenance: Add `@return-taint none` to getArg and getOption
In a generic context, $argv would be considered tainted by taint-check,
which makes sense for certain applications. In the context of MW
maintenance scripts, however, we can assume that arguments are trusted,
because maintenance scripts can only be executed by sysadmins, who are
generally trusted (and even if they aren't, they have other ways to
compromise a wiki).

Bug: T216348
Change-Id: I8e171f5dcd5b9232d45e85d775cd794f2ebb2dfc